+++ go.mod
@@ -1,1 +1,6 @@
 module github.com/renderedtext/agent
+
+require (
+	github.com/ghodss/yaml v1.0.0
+	gopkg.in/yaml.v2 v2.2.2
+)
